Hook up GN official build and branding.
This passes the official build and branding flags to the GN build. Currently, this is used to hook up the proper sysroot for compiling on Linux for official builds. The 32-bit branded build does not compile currently due to pending pkgconfig issues.

Fix a bug where assertions in the build config would not get reported properly. This was because I forgot to actually throw the error at the toplevel.

BUG=
R=piman@chromium.org

Review URL: https://codereview.chromium.org/115643002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 42b013fd9d2870c19d16fb4a1391b2343d29c7d1