Require commas between items in a GN list.
Fixing this issue exposed a race condition in .gni processing where an error parsing a file that another thread was waiting for wouldn't signal that other thread that the load was complete, so it would wait forever.

This patch separates out the loading from the signaling in the InputFileManager so we can't forget to do this.

BUG=368019
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/258073004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7d9b35982ebc5ae46c553f0569f45cfe29f7b86